Moved helper functions to utils.py
