+++ src/soa_builder/web/utils.py
@@ -47,16 +47,21 @@ _ISO_DURATION_RE = re.compile(
 )
 
 
-# Helper function to convert ISO-8601 duration/period strings
-# to days, using these common approximations for years and months
-"""
-    1 year = 365 days
-    1 month = 30 days
-    1 week = 7 days
-    1 hour = 1/24 day
-    1 minute = 1/(24*60) day
-    1 second = 1/(24*3600) day
-"""
+# USDM JSON generator helper
+def _get_biomedical_concept_ids(soa_id: int, activity_uid: int) -> List[str]:
+    conn = _connect()
+    cur = conn.cursor()
+    cur.execute(
+        "SELECT concept_uid from activity_concept where soa_id=? and activity_uid=?",
+        (
+            soa_id,
+            activity_uid,
+        ),
+    )
+    rows = cur.fetchall()
+    conn.close()
+    bc_uids = [r[0] for r in rows] or []
+    return bc_uids
 
 
 def redirect_url_from_referer(request: Request, fallback: str) -> str:
@@ -68,6 +73,18 @@ def redirect_url_from_referer(request: Request, fallback: str) -> str:
         if parsed.netloc == base.netloc and parsed.path.startswith("/ui/"):
             return urlunparse(("", "", parsed.path, "", parsed.query, parsed.fragment))
     return fallback
+
+
+# Helper function to convert ISO-8601 duration/period strings
+# to days, using these common approximations for years and months
+"""
+    1 year = 365 days
+    1 month = 30 days
+    1 week = 7 days
+    1 hour = 1/24 day
+    1 minute = 1/(24*60) day
+    1 second = 1/(24*3600) day
+"""
 
 
 def iso_duration_to_days(iso_duration: str) -> float:
